adequei as classes Inventario e Item, para configurar uma composição, também exclui a classe Produto
